🔑 added Yubikey KeychainViewCell

+++ PGPro/Keychain/Controller/KeychainViewController.swift
@@ -18,6 +18,7 @@
 import UIKit
 import MobileCoreServices
 import ObjectivePGP
+import YubiKit
 
 class KeychainViewController: UIViewController {
     
@@ -242,34 +243,56 @@ extension KeychainViewController: UIDocumentPickerDelegate {
 // MARK: - Table View
 
 extension KeychainViewController: UITableViewDataSource, UITableViewDelegate {
+
+    var yubikey: Int {
+        return (Preferences.yubikey ? 1 : 0)
+    }
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         if isFiltering() {
             return filteredContacts.count
         } else {
-            return contacts.count
+            return contacts.count + yubikey
         }
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        var cntct = contacts[indexPath.row]
+        if (!isFiltering() && indexPath.row == 0 && Preferences.yubikey) {
+            return YubikeyTableViewCell()
+        }
+        var contactIndex = indexPath.row - yubikey
         if isFiltering() {
-            cntct = filteredContacts[indexPath.row]
+            contactIndex += yubikey
+        }
+
+        var cntct = contacts[contactIndex]
+        if isFiltering() {
+            cntct = filteredContacts[contactIndex]
         }
         if let cell = keychainTableView.dequeueReusableCell(withIdentifier: "KeychainTableViewCell") as? KeychainTableViewCell {
             cell.setContact(contact: cntct)
             return cell
         } else {
-            return UITableViewCell() // Dummy return value
+            return UITableViewCell()
         }
     }
     
     func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
+        if (!isFiltering() && indexPath.row == 0 && Preferences.yubikey) {
+            Log.s("Can't delete the Yubikey cell!")
+            return // Don't delete the Yubikey cell
+        }
+
+        var contactIndex = indexPath.row - yubikey
+        if isFiltering() {
+            contactIndex += yubikey
+        }
+
         if (editingStyle == .delete) {
             if isFiltering() {
-                let cntct = filteredContacts[indexPath.row]
+                let cntct = filteredContacts[contactIndex]
 
-                filteredContacts.remove(at: indexPath.row)
+                filteredContacts.remove(at: contactIndex)
                 keychainTableView.deleteRows(at: [indexPath], with: .bottom)
 
                 ContactListService.remove(cntct)
@@ -277,7 +300,7 @@ extension KeychainViewController: UITableViewDataSource, UITableViewDelegate {
                 
             } else {
                 // Remove from storage and update local list
-                ContactListService.remove(contacts[indexPath.row])
+                ContactListService.remove(contacts[contactIndex])
                 contacts = ContactListService.get(ofType: .both)
 
                 // Remove from view and update view
@@ -287,9 +310,19 @@ extension KeychainViewController: UITableViewDataSource, UITableViewDelegate {
     }
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        var contact = contacts[indexPath.row]
+        if (!isFiltering() && indexPath.row == 0 && Preferences.yubikey) {
+            // TODO: This code gets executed when the Yubikey button is pressed
+            return
+        }
+
+        var contactIndex = indexPath.row - yubikey
         if isFiltering() {
-            contact = filteredContacts[indexPath.row]
+            contactIndex += yubikey
+        }
+
+        var contact = contacts[contactIndex]
+        if isFiltering() {
+            contact = filteredContacts[contactIndex]
         }
 
         let detailViewController = ContactDetailViewController()
